requirements.txt
requests
vertexai
google-cloud-texttospeech
PyYAML